added: bounds class

+++ src/ray.h
@@ -11,7 +11,7 @@ private:
   mutable float tMax;
   float time;
 public:
-  Ray(Vector3 origin, Vector3 direction, float tMax = std::numeric_limits<int>::max(), float time = 0.0f);
+  Ray(Vector3 origin, Vector3 direction, float tMax = std::numeric_limits<float>::max(), float time = 0.0f);
   ~Ray();
 
   Vector3 operator()(float t);
